Clinica_OdontologicaW/src/main/java/com/backend/Clinica_OdontologicaW/service/impl/PacienteService.java
package com.backend.Clinica_OdontologicaW.service.impl;

import com.backend.Clinica_OdontologicaW.dao.IDao;
import com.backend.Clinica_OdontologicaW.model.Paciente;
import com.backend.Clinica_OdontologicaW.service.IPacienteService;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class PacienteService implements IPacienteService {
    private IDao<Paciente> pacienteIDao;
    public PacienteService(IDao<Paciente> pacienteIDao) {
        this.pacienteIDao = pacienteIDao;
    }
    @Override
    public Paciente registrarPaciente(Paciente paciente) {
        return pacienteIDao.registrar(paciente);
    }

    @Override
    public List<Paciente> listarPacientes() {
        return pacienteIDao.listarTodos();
    }

    @Override
    public Paciente buscarPacientePorId(int id) {
        return pacienteIDao.buscarPorId(id);
    }

    @Override
    public Paciente actualizarPaciente(Paciente paciente) {
        return pacienteIDao.actualizar(paciente);
    }
}
